apiv2/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.source.txt
#
annotated-types==0.6.0
    # via pydantic
anyio==4.0.0
    # via
    #   httpx
    #   watchfiles
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
dnspython==2.4.2
    # via email-validator
email-validator==2.1.0.post1
    # via pydantic
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.25.1
    # via -r requirements.source.txt
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.2
    # via -r requirements.source.txt
loguru==0.7.2
    # via -r requirements.source.txt
markupsafe==2.1.3
    # via jinja2
pampy==0.3.0
    # via -r requirements.source.txt
psycopg[binary,pool]==3.1.12
    # via -r requirements.source.txt
psycopg-binary==3.1.12
    # via psycopg
psycopg-pool==3.2.0
    # via psycopg
pydantic[email]==2.5.1
    # via -r requirements.source.txt
pydantic-core==2.14.3
    # via pydantic
pyjwt==2.8.0
    # via -r requirements.source.txt
python-decouple==3.8
    # via -r requirements.source.txt
python-dotenv==1.0.0
    # via uvicorn
pyyaml==6.0.1
    # via uvicorn
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
typing-extensions==4.8.0
    # via
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
uvicorn[standard]==0.24.0.post1
    # via -r requirements.source.txt
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
